+++ fsae-vehicle-fw/src/main.cpp
@@ -7,16 +7,16 @@
 #include "peripherals/can.h"
 
 #include "peripherals/gpio.h"
-#include "vehicle/apps.h"
-#include "vehicle/bse.h"
-#include "vehicle/can_polls.h"
+#include "vehicle/comms/can_polls.h"
+#include "vehicle/comms/pcc_receive.h"
+#include "vehicle/devices/apps.h"
+#include "vehicle/devices/bse.h"
+#include "vehicle/devices/rtm_button.h"
+#include "vehicle/devices/shockTravel.h"
+#include "vehicle/devices/thermal.h"
 #include "vehicle/faults.h"
 #include "vehicle/motor.h"
-#include "vehicle/pcc_receive.h"
-#include "vehicle/rtm_button.h"
-#include "vehicle/shockTravel.h"
 #include "vehicle/telemetry.h"
-#include "vehicle/thermal.h"
 
 #include "utils/utils.h"
 #include <iostream>
