+++ fsae-vehicle-fw/src/vehicle/devices/apps.h
