+++ fsae-vehicle-fw/src/vehicle/devices/bse.h
