+++ fsae-vehicle-fw/src/vehicle/devices/shockTravel.h
