+++ fsae-vehicle-fw/src/vehicle/devices/thermal.cpp
